backend-21/pretups/src/main/java/com/restapi/superadmin/responseVO/AddCategoryResponseVO.java
package com.restapi.superadmin.responseVO;

import com.btsl.common.BaseResponse;

public class AddCategoryResponseVO extends BaseResponse {
	
	private String categoryCode;
	private String domainCode;
	private String agentAllowed;

	public String getDomainCode() {
		return domainCode;
	}

	public void setDomainCode(String domainCode) {
		this.domainCode = domainCode;
	}

	public String getAgentAllowed() {
		return agentAllowed;
	}

	public void setAgentAllowed(String agentAllowed) {
		this.agentAllowed = agentAllowed;
	}

	public String getCategoryCode() {
		return categoryCode;
	}

	public void setCategoryCode(String categoryCode) {
		this.categoryCode = categoryCode;
	}
	
	

}
